requirements.txt
fitbit==0.0.2
django-celery==3.0.17
